+++ findCandidates.h
@@ -1,4 +1,4 @@
-struct orderedList calcCandidates(double* peaks, int numPeaks);
+struct orderedList calcCandidates(float* peaks, int numPeaks);
 float calcM(float f_i, float f_j);
 struct distinctList* distinctCandidates(struct orderedList* candidates,
 					int max_length, float xi,
